requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
bitstring==3.1.6
boto3==1.12.5
botocore==1.15.5
certifi==2019.6.16
cffi==1.12.3
cfn-flip==1.2.2
chardet==3.0.4
Click==7.0
codacy-coverage==1.3.11
coverage==4.5.3
cryptography==2.7
dnslib==0.9.7
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
flake8==3.7.8
flake8-html==0.4.0
idna==2.8
importlib-metadata==0.18
isort==4.3.21
Jinja2==2.10.1
jmespath==0.9.4
katnip==0.2.5
kittyfuzzer==0.7.4
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mock==3.0.5
more-itertools==7.2.0
packaging==19.0
pipenv==2018.11.26
pluggy==0.13.1
prettytable==0.7.2
protobuf==3.13.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pytest==5.1.3
python-dateutil==2.8.1
PyYAML==5.2
redis==3.3.11
requests==2.22.0
s3transfer==0.3.3
six==1.15.0
toml==0.10.0
tox==3.20.0
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.6.1
virtualenv-clone==0.5.3
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.5.2
